+++ Magic/CHANGELOG.md
@@ -6,6 +6,7 @@
  - Fix Cleric chestplate preventing Recall
  - Fixes and improvements to "mmob clear" and "mmob list" commands
  - Allow protected and powered properties on classes
+ - Add max_item_count parameter to GiveItem action, for filling a player's item up to a certain amount
 
 # 7.9.13